Spike relocating event comments to event scraper
